Доработка проекта
