+++ README.md
@@ -1,26 +1,17 @@
-# DiplomFA
-Базы данных:
+#DiplomFA
 
-Задание 1
-Проверка, отображается ли созданный заказ в базе данных.
+#Работа с базой данных Яндекс Самокат:
+
+1. SQL-запрос в консоли проверяет, отображается ли созданный заказ в базе данных.
 Для этого выведен список логинов курьеров с количеством их заказов в статусе «В доставке» (поле inDelivery = true). 
 
-Задание 2
-Тест статусов заказов. Нужно убедиться, что в базе данных они записываются корректно.
+2. SQL-запрос для тестирования статусов заказов. 
+Нужно убедиться, что в базе данных они записываются корректно.
 Для этого выведены все трекеры заказов и их статусы. 
-Статусы определяются по следующему правилу:
-Если поле finished == true, то выводится статус 2.
-Если поле canсelled == true, то выводится статус -1.
-Если поле inDelivery == true, то выводится статус 1.
-Для остальных случаев выводится 0.
 
-Автоматизация:
-Автоматизация сценария:
-1. Клиент создает заказ.
-2. Проверяется, что по треку заказа можно получить данные о заказе.
+#Автоматизация тестирования Яндекс Самокат:
+Автоматизация сценария: проверяется, что по треку заказа можно получить данные о созданном заказе.
 
-Шаги автотеста:
-1.Выполнить запрос на создание заказа.
-2. Сохранить номер трека заказа.
-3. Выполнить запрос на получения заказа по треку заказа.
-4. Проверить, что код ответа равен 200.
+#Требования для проведения автоматизированного тестирования: 
+- Для запуска теста должны быть установлены пакеты pytest и requests
+- Запуск теста выполняется командой pytest autotest.py
