+++ configuration.py
@@ -1,4 +1,5 @@
 # Федоренко Антон, 25-ая когорта, Дипломный проект. Инженер по тестированию плюс
 
-URL_SERVICE = "https://0511f7f5-fb88-43c3-82f8-1a91708e4dcd.serverhub.praktikum-services.ru/"
+URL_SERVICE = "https://94fab5b2-b579-424b-8f63-66408b219878.serverhub.praktikum-services.ru"
 CREAT_ORDERS = "/api/v1/orders"
+
